Updated CSS - added new classes and more specific selectors
Used the new classes that were reworked to evenly space out the receipes with images on top and links on bottom.
Changed background color to match link and title color to match the same color palette.
Used inline block for receipe div to have all elements sit next to each other and sit on one row.